My contribution to the VectorNav Data Stream. I have a few suggestions for the VectorNav SDK when trying to access the CompositeData.
When I program the SDK, I notice I have to call each field, which while efficient for the program, makes the programmer's life a little difficult to code each measurement type and record the appropriate data. I was thinking about a type of Multipart DataType scheme that would allow you to use a method that takes in the measuregroup type and measurefield type that then calls that field automatically and returns it to the user. That would help reduce the amount of lines in the code and make the file size shorter.

Another suggestion might be that at runtime, we could make a Queue of recorded data types passed from the sensor to the SDK that then the programmer could invoke and with the method idea, could allow readability to port over the necessary values.

Thank you for this project and I hope to have similar opportunities in the future!